Add Scheduler defaults to ModelTemplate
