+++ OnnxStack.UI/Models/StableDiffusionModelTemplate.cs
@@ -1,0 +1,9 @@
+﻿using OnnxStack.StableDiffusion.Enums;
+
+namespace OnnxStack.UI.Models
+{
+    public record StableDiffusionModelTemplate(DiffuserPipelineType PipelineType, ModelType ModelType, int SampleSize, DiffuserType[] DiffuserTypes)
+    {
+        public StableDiffusionSchedulerDefaults SchedulerDefaults { get; set; } = new StableDiffusionSchedulerDefaults();
+    }
+}
