+++ OnnxStack.UI/UserControls/SchedulerControl.xaml
@@ -163,7 +163,7 @@
                                 <Label>Inference Steps</Label>
                                 <TextBlock Text="{Binding ElementName=SliderInferenceSteps, Path=Value, StringFormat={}{0}}" VerticalAlignment="Bottom" HorizontalAlignment="Right" FontSize="10" Margin="0,0,6,0" FontWeight="Medium" />
                             </DockPanel>
-                            <Slider Name="SliderInferenceSteps" Value="{Binding SchedulerOptions.InferenceSteps}" Minimum="1" Maximum="{Binding SchedulerOptions.OriginalInferenceSteps}" TickFrequency="1" IsSnapToTickEnabled="True" SmallChange="1" LargeChange="1" >
+                            <Slider Name="SliderInferenceSteps" Value="{Binding SchedulerOptions.InferenceSteps}" Minimum="{Binding SchedulerDefaults.StepsMin}" Maximum="{Binding SchedulerDefaults.StepsMax}" TickFrequency="1" IsSnapToTickEnabled="True" SmallChange="1" LargeChange="1" >
                                 <i:Interaction.Behaviors>
                                     <behaviors:SliderMouseWheelBehavior />
                                 </i:Interaction.Behaviors>
@@ -174,7 +174,7 @@
                                 <Label>Guidance Scale</Label>
                                 <TextBlock Text="{Binding ElementName=SliderGuidanceScale, Path=Value, StringFormat={}{0:F2}}" VerticalAlignment="Bottom" HorizontalAlignment="Right" FontSize="10" Margin="0,0,6,0" FontWeight="Medium" />
                             </DockPanel>
-                            <Slider Name="SliderGuidanceScale" Value="{Binding SchedulerOptions.GuidanceScale}" Minimum="0" Maximum="30" TickFrequency="0.1" IsSnapToTickEnabled="True" SmallChange="0.1" LargeChange="0.1">
+                            <Slider Name="SliderGuidanceScale" Value="{Binding SchedulerOptions.GuidanceScale}" Minimum="{Binding SchedulerDefaults.GuidanceMin}" Maximum="{Binding SchedulerDefaults.GuidanceMax}" TickFrequency="0.1" IsSnapToTickEnabled="True" SmallChange="0.1" LargeChange="0.1">
                                 <i:Interaction.Behaviors>
                                     <behaviors:SliderMouseWheelBehavior />
                                 </i:Interaction.Behaviors>
