+++ OnnxStack.UI/appsettings.json
@@ -49,7 +49,7 @@
       {
         "Name": "StableDiffusion",
         "ImageIcon": "",
-        "Rank":  100,
+        "Rank": 100,
         "Author": "OnnxStack",
         "Category": "StableDiffusion",
         "Template": "StableDiffusion",
@@ -61,7 +61,10 @@
             "TextToImage",
             "ImageToImage",
             "ImageInpaintLegacy"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "SchedulerType": "EulerAncestral"
+          }
         }
       },
       {
@@ -72,11 +75,15 @@
         "Category": "StableDiffusion",
         "Template": "StableDiffusionInpaint",
         "StableDiffusionTemplate": {
+          "SampleSize": 512,
           "ModelType": "Base",
           "PipelineType": "StableDiffusion",
           "DiffuserTypes": [
             "ImageInpaint"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "SchedulerType": "EulerAncestral"
+          }
         }
       },
       {
@@ -87,13 +94,17 @@
         "Category": "StableDiffusion",
         "Template": "StableDiffusionXL",
         "StableDiffusionTemplate": {
+          "SampleSize": 1024,
           "ModelType": "Base",
           "PipelineType": "StableDiffusionXL",
           "DiffuserTypes": [
             "TextToImage",
             "ImageToImage",
             "ImageInpaintLegacy"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "SchedulerType": "EulerAncestral"
+          }
         }
       },
       {
@@ -104,11 +115,15 @@
         "Category": "StableDiffusion",
         "Template": "StableDiffusionXLInpaint",
         "StableDiffusionTemplate": {
+          "SampleSize": 1024,
           "ModelType": "Base",
           "PipelineType": "StableDiffusionXL",
           "DiffuserTypes": [
             "ImageInpaint"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "SchedulerType": "EulerAncestral"
+          }
         }
       },
       {
@@ -119,13 +134,43 @@
         "Category": "StableDiffusion",
         "Template": "StableDiffusionXLRefiner",
         "StableDiffusionTemplate": {
+          "SampleSize": 1024,
+          "ModelType": "Refiner",
+          "PipelineType": "StableDiffusionXL",
+          "DiffuserTypes": [
+            "TextToImage",
+            "ImageToImage",
+            "ImageInpaintLegacy"
+          ],
+          "SchedulerDefaults": {
+            "SchedulerType": "EulerAncestral"
+          }
+        }
+      },
+      {
+        "Name": "StableDiffusion XL Turbo",
+        "ImageIcon": "",
+        "Rank": 0,
+        "Author": "OnnxStack",
+        "Category": "StableDiffusion",
+        "Template": "StableDiffusionXLTurbo",
+        "StableDiffusionTemplate": {
+          "SampleSize": 512,
           "ModelType": "Base",
           "PipelineType": "StableDiffusionXL",
           "DiffuserTypes": [
             "TextToImage",
             "ImageToImage",
             "ImageInpaintLegacy"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "Steps": 4,
+            "StepsMin": 1,
+            "Guidance": 0,
+            "GuidanceMin": 0,
+            "GuidanceMax": 0,
+            "SchedulerType": "EulerAncestral"
+          }
         }
       },
       {
@@ -136,13 +181,21 @@
         "Category": "StableDiffusion",
         "Template": "LatentConsistency",
         "StableDiffusionTemplate": {
+          "SampleSize": 512,
           "ModelType": "Base",
           "PipelineType": "LatentConsistency",
           "DiffuserTypes": [
             "TextToImage",
             "ImageToImage",
             "ImageInpaintLegacy"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "Steps": 6,
+            "Guidance": 1,
+            "GuidanceMin": 0,
+            "GuidanceMax": 2,
+            "SchedulerType": "LCM"
+          }
         }
       },
       {
@@ -153,13 +206,21 @@
         "Category": "StableDiffusion",
         "Template": "LatentConsistencyXL",
         "StableDiffusionTemplate": {
+          "SampleSize": 1024,
           "ModelType": "Base",
           "PipelineType": "LatentConsistencyXL",
           "DiffuserTypes": [
             "TextToImage",
             "ImageToImage",
             "ImageInpaintLegacy"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "Steps": 6,
+            "Guidance": 1,
+            "GuidanceMin": 0,
+            "GuidanceMax": 2,
+            "SchedulerType": "LCM"
+          }
         }
       },
       {
@@ -170,11 +231,21 @@
         "Category": "StableDiffusion",
         "Template": "InstaFlow",
         "StableDiffusionTemplate": {
+          "SampleSize": 512,
           "ModelType": "Base",
           "PipelineType": "InstaFlow",
           "DiffuserTypes": [
             "TextToImage"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "Steps": 1,
+            "StepsMin": 1,
+            "StepsMax": 1,
+            "Guidance": 0,
+            "GuidanceMin": 0,
+            "GuidanceMax": 0,
+            "SchedulerType": "InstaFlow"
+          }
         }
       },
       {
@@ -205,7 +276,10 @@
             "TextToImage",
             "ImageToImage",
             "ImageInpaintLegacy"
-          ]
+          ],
+          "SchedulerDefaults": {
+            "SchedulerType": "EulerAncestral"
+          }
         },
         "Repository": "https://huggingface.co/runwayml/stable-diffusion-v1-5",
         "RepositoryBranch": "onnx",
